--- burger-queen-project/src/pages/components/boxBurgerDance.js
@@ -1,8 +1,0 @@
-export function BoXBurgerDance () {
-    return (
-        <div className="burgerBox">
-            <h1>BURGER DANCING</h1>
-        </div>
-        
-    )
-  }
