--- burger-queen-project/src/pages/components/headerMenu.js
@@ -1,9 +1,0 @@
-export function HeaderMenu () {
-    return (
-        <div className="header">
-          <header className="header">
-            <h1>BURGER</h1>
-          </header>
-        </div>
-    )
-  }
