+++ burger-queen-project/src/styles/globals.css
@@ -3,9 +3,7 @@
 @tailwind utilities; */
 
 :root {
-  /* --foreground-rgb: 0, 0, 0;
-  --background-start-rgb: 214, 219, 220;
-  --background-end-rgb: 255, 255, 255; */
+  --button-background: yellow;
 }
 
 @media (prefers-color-scheme: dark) {
@@ -13,6 +11,7 @@
     --foreground-rgb: 255, 255, 255;
     --background-start-rgb: 0, 0, 0;
     --background-end-rgb: 0, 0, 0;
+    --button-background: yellow;
   }
 }
