+++ burger-queen-project/src/styles/headerMenu.module.css
@@ -1,0 +1,10 @@
+/*Component header*/
+.logoBurger{
+    width: 50px;
+    height: 50px;
+    border-radius: 200%;
+    background: black;
+  }
+  .btnnn{
+    background: var(--button-background);
+  }
